adattamento codice in concomitanza con front end angular: sviluppato controller advice, perfezionati i DTO e bug fix varie.

+++ src/main/java/it/trainingsi2001/rentalcars/service/TipologiaServiceImpl.java
@@ -3,18 +3,23 @@ package it.trainingsi2001.rentalcars.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import it.trainingsi2001.rentalcars.entities.Tipologia;
 import it.trainingsi2001.rentalcars.repository.TipologiaRepository;
+import it.trainingsi2001.rentalcars.repository.VeicoloRepository;
 
 @Service
 @Transactional(readOnly = true)
-public class TipologiaServiceImpl implements TipologiaService{
+public class TipologiaServiceImpl implements TipologiaService {
 
     @Autowired
     TipologiaRepository tipologiaRepository;
+
+    @Autowired
+    VeicoloRepository veicoloRepository;
 
     @Override
     public List<Tipologia> listAll() {
@@ -41,6 +46,9 @@ public class TipologiaServiceImpl implements TipologiaService{
     @Override
     @Transactional
     public void rimuovi(Long idToRemove) {
+        if (veicoloRepository.existsByTipologia_Id(idToRemove)) {
+            throw new IllegalStateException("Impossibile eliminare: ci sono veicoli associati a questa tipologia.");
+        }
         tipologiaRepository.deleteById(idToRemove);
     }
